pyproject.toml
[project]
name = "metal-live-scene-analytics"
version = "0.1.0"
description = "Metal Live Scene Analytics — DE Zoomcamp Capstone"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-bigquery>=1.9.0",
    "dbt-core>=1.9.0",
    "google-cloud-bigquery>=3.25",
    "google-cloud-storage>=2.18",
    "pandas>=2.0",
    "plotly>=6.0",
    "prefect>=3.0",
    "kaggle>=2.0.0",
    "requests>=2.32.5",
    "streamlit>=1.40",
]

[dependency-groups]
dev = [
    "google-cloud-bigquery-storage>=2.36.2",
]
